fix: move logger setup to __init__
